+++ cmd/psweb/ln/common.go
@@ -301,28 +301,30 @@ func OnMyCustomMessage(nodeId string, payload []byte) {
 		// repeat invite to ClaimJoin
 		shareInvite(nodeId)
 
-		// repeat last
-		if AdvertiseLiquidBalance && SentLiquidBalances[nodeId] != nil {
-			if SendCustomMessage(nodeId, &Message{
-				Version: MESSAGE_VERSION,
-				Memo:    "balance",
-				Asset:   "lbtc",
-				Amount:  SentLiquidBalances[nodeId].Amount,
-			}) == nil {
-				// save timestamp
-				SentLiquidBalances[nodeId].TimeStamp = time.Now().Unix()
+		if config.Config.AllowSwapRequests {
+			// repeat last
+			if AdvertiseLiquidBalance && SentLiquidBalances[nodeId] != nil {
+				if SendCustomMessage(nodeId, &Message{
+					Version: MESSAGE_VERSION,
+					Memo:    "balance",
+					Asset:   "lbtc",
+					Amount:  SentLiquidBalances[nodeId].Amount,
+				}) == nil {
+					// save timestamp
+					SentLiquidBalances[nodeId].TimeStamp = time.Now().Unix()
+				}
 			}
-		}
 
-		if AdvertiseBitcoinBalance && SentBitcoinBalances[nodeId] != nil {
-			if SendCustomMessage(nodeId, &Message{
-				Version: MESSAGE_VERSION,
-				Memo:    "balance",
-				Asset:   "btc",
-				Amount:  SentBitcoinBalances[nodeId].Amount,
-			}) == nil {
-				// save timestamp
-				SentBitcoinBalances[nodeId].TimeStamp = time.Now().Unix()
+			if config.Config.BitcoinSwaps && AdvertiseBitcoinBalance && SentBitcoinBalances[nodeId] != nil {
+				if SendCustomMessage(nodeId, &Message{
+					Version: MESSAGE_VERSION,
+					Memo:    "balance",
+					Asset:   "btc",
+					Amount:  SentBitcoinBalances[nodeId].Amount,
+				}) == nil {
+					// save timestamp
+					SentBitcoinBalances[nodeId].TimeStamp = time.Now().Unix()
+				}
 			}
 		}
 
@@ -524,11 +526,7 @@ func DecodeAndProcessInvoice(bolt11 string, valueMsat int64) bool {
 	}
 
 	// Decode the payment request
-	var harnessNetParams = &chaincfg.MainNetParams
-	if config.Config.Chain == "testnet" {
-		harnessNetParams = &chaincfg.TestNet3Params
-	}
-	invoice, err := zpay32.Decode(bolt11, harnessNetParams)
+	invoice, err := zpay32.Decode(bolt11, getHarnessNetParams())
 
 	if err == nil {
 		if invoice.Description != nil {
@@ -574,4 +572,22 @@ func lastFeeIsTheSame(channelId uint64, newFee int, isInbound bool) bool {
 		}
 	}
 	return false
+}
+
+func getHarnessNetParams() *chaincfg.Params {
+	switch config.Config.Chain {
+	case "regtest":
+		return &chaincfg.RegressionNetParams
+	case "testnet":
+		return &chaincfg.TestNet3Params
+	case "testnet4":
+		return &chaincfg.TestNet4Params
+	case "signet":
+		return &chaincfg.SigNetParams
+	case "mainnet":
+		return &chaincfg.MainNetParams
+	}
+
+	log.Panicf("Chain %s is not supported!")
+	return nil
 }