+++ cmd/psweb/ln/lnd.go
@@ -29,7 +29,6 @@ import (
 
 	"github.com/btcsuite/btcd/btcutil"
 	"github.com/btcsuite/btcd/btcutil/psbt"
-	"github.com/btcsuite/btcd/chaincfg"
 	"github.com/btcsuite/btcd/txscript"
 	"github.com/btcsuite/btcd/wire"
 
@@ -129,14 +128,10 @@ func lndConnection() (*grpc.ClientConn, error) {
 
 	opts := []grpc.DialOption{
 		grpc.WithTransportCredentials(tlsCreds),
-		grpc.WithBlock(),
 		grpc.WithPerRPCCredentials(macCred),
 	}
 
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	conn, err := grpc.DialContext(ctx, host, opts...)
+	conn, err := grpc.NewClient(host, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -538,12 +533,7 @@ func fundPsbtSpendAll(cl walletrpc.WalletKitClient, utxoStrings *[]string, addre
 		}
 	}
 
-	var harnessNetParams = &chaincfg.TestNet3Params
-	if config.Config.Chain == "mainnet" {
-		harnessNetParams = &chaincfg.MainNetParams
-	}
-
-	parsed, err := btcutil.DecodeAddress(address, harnessNetParams)
+	parsed, err := btcutil.DecodeAddress(address, getHarnessNetParams())
 	if err != nil {
 		log.Println("DecodeAddress:", err)
 		return nil, err